requirements.txt
fastapi
uvicorn
pandas
python-multipart
openpyxl
unidecode
scikit-learn
numpy
python-dateutil